refactor: MapStruct 매핑 시 unmapped 경고 무시하도록 설정 변경 (#79)
