+++ src/main/java/com/harusari/chainware/member/common/mapstruct/MemberMapStruct.java
@@ -3,8 +3,12 @@ package com.harusari.chainware.member.common.mapstruct;
 import com.harusari.chainware.member.command.application.dto.request.MemberCreateRequest;
 import com.harusari.chainware.member.command.domain.aggregate.Member;
 import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
 
-@Mapper
+@Mapper(
+        unmappedSourcePolicy =  ReportingPolicy.IGNORE,
+        unmappedTargetPolicy = ReportingPolicy.IGNORE
+)
 public interface MemberMapStruct {
 
     Member toMember(MemberCreateRequest memberCreateRequest);
